# target/generated-sources/openapi/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: backend@sii.com
  description: Next Catalog Service API (CATAS)
  title: next-catalog-service-api
  version: "0.5"
servers:
- url: http://tbd
...
paths:
  /catalog/{carlineId}/modelId}:
    get:
      description: Return the entire catalog for a model.
      operationId: getCatalogDataForModel
      parameters:
      - explode: false
        in: path
        name: carlineId
        required: true
        schema:
          description: The id of the carline
          example: a4limo
          type: string
        style: simple
      - explode: true
        in: query
        name: modelId
        required: true
        schema:
          description: The id of the model
          example: 8vfanc-0-2020
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
          description: Returning a list of exterior
          headers:
            X-DataVersion:
              description: Data version
              example: mv-3131
              explode: false
              schema:
                type: string
              style: simple
        "401":
          description: Authentication information is missing or invalid
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "404":
          description: The specified resource was not found.
      summary: List all equipments by model id.
      tags:
      - next-catalog-data
      x-accepts: application/json
      x-tags:
      - tag: next-catalog-data
components:
...
  schemas:
    CatalogResponse:
      example:
        specialGroups:
        - equipmentFamilies:
          - equipments:
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            name: Exterior
            id: exterior.exterior
          - equipments:
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            name: Exterior
            id: exterior.exterior
          name: Infotainment/Technology
          id: infotainment
        - equipmentFamilies:
          - equipments:
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            name: Exterior
            id: exterior.exterior
          - equipments:
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            name: Exterior
            id: exterior.exterior
          name: Infotainment/Technology
          id: infotainment
        extColors:
        - familyName: Uni-Lacke
          id: color-type:uni
          equipmentList:
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
        - familyName: Uni-Lacke
          id: color-type:uni
          equipmentList:
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
        carlineId: a1sb
        bodyColors:
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
        roofColors:
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
        accesories:
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
        stealthItems:
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
        groups:
        - equipmentFamilies:
          - equipments:
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            name: Exterior
            id: exterior.exterior
          - equipments:
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            name: Exterior
            id: exterior.exterior
          name: Infotainment/Technology
          id: infotainment
        - equipmentFamilies:
          - equipments:
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            name: Exterior
            id: exterior.exterior
          - equipments:
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            - mbvId: MRADC2D
              texts:
              - null
              - null
              price: 0.0
              renderCode: MRADC2D
              name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
              attributes:
                key: attributes
              id: C2D
              isVisible: true
              marginable: true
              media:
              - null
              - null
              equipmentType: option
              content:
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
              - mbvId: tbd
                name: tbd
                evaKey: KK1
                id: tbd
            name: Exterior
            id: exterior.exterior
          name: Infotainment/Technology
          id: infotainment
        interiorTiles:
        - familyName: Uni-Lacke
          id: color-type:uni
          equipmentList:
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
        - familyName: Uni-Lacke
          id: color-type:uni
          equipmentList:
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
      properties:
        carlineId:
          example: a1sb
          type: string
        groups:
          items:
            $ref: '#/components/schemas/EquipmentGroup'
          type: array
        extColors:
          items:
            $ref: '#/components/schemas/ExteriorColor'
          type: array
        interiorTiles:
          items:
            $ref: '#/components/schemas/InteriorTiles'
          type: array
        specialGroups:
          items:
            $ref: '#/components/schemas/EquipmentGroup'
          type: array
        bodyColors:
          items:
            $ref: '#/components/schemas/Equipment'
          type: array
        roofColors:
          items:
            $ref: '#/components/schemas/Equipment'
          type: array
        stealthItems:
          items:
            $ref: '#/components/schemas/Equipment'
          type: array
        accesories:
          items:
            $ref: '#/components/schemas/Equipment'
          type: array
      type: object
    EquipmentGroup:
      example:
        equipmentFamilies:
        - equipments:
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
          name: Exterior
          id: exterior.exterior
        - equipments:
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
          - mbvId: MRADC2D
            texts:
            - null
            - null
            price: 0.0
            renderCode: MRADC2D
            name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
            attributes:
              key: attributes
            id: C2D
            isVisible: true
            marginable: true
            media:
            - null
            - null
            equipmentType: option
            content:
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
            - mbvId: tbd
              name: tbd
              evaKey: KK1
              id: tbd
          name: Exterior
          id: exterior.exterior
        name: Infotainment/Technology
        id: infotainment
      properties:
        id:
          example: infotainment
          type: string
        name:
          description: Name of the mbv lasche
          example: Infotainment/Technology
          type: string
        equipmentFamilies:
          items:
            $ref: '#/components/schemas/EquipmentFamily'
          type: array
      type: object
    ExteriorColor:
      example:
        familyName: Uni-Lacke
        id: color-type:uni
        equipmentList:
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
      properties:
        id:
          example: color-type:uni
          type: string
        familyName:
          example: Uni-Lacke
          type: string
        equipmentList:
          items:
            $ref: '#/components/schemas/Equipment'
          type: array
      type: object
    InteriorTiles:
      example:
        familyName: Uni-Lacke
        id: color-type:uni
        equipmentList:
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
        - mbvId: MRADC2D
          texts:
          - null
          - null
          price: 0.0
          renderCode: MRADC2D
          name: "Räder, 10-Speichen, 6,5Jx16, Reifen 195/55 R16"
          attributes:
            key: attributes
          id: C2D
          isVisible: true
          marginable: true
          media:
          - null
          - null
          equipmentType: option
          content:
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
          - mbvId: tbd
            name: tbd
            evaKey: KK1
            id: tbd
      properties:
        id:
          example: color-type:uni
          type: string
        familyName:
          example: Uni-Lacke
          type: string
        equipmentList:
          items:
            $ref: '#/components/schemas/Equipment'
          type: array
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-KEY
      type: apiKey